+++ pointofsale/Product.cs
@@ -8,11 +8,11 @@ namespace pointofsale
 {
     class Product
     {
-        public double price;
+        public decimal price;
         public string name;
 
 
-        public Product(double price, string name)
+        public Product(decimal price, string name)
         {
             this.price = price;
             this.name = name;
