+++ pointofsale/ProductArray.cs
@@ -13,16 +13,16 @@ namespace pointofsale
         public Product[] ReturnArray()
         {
             Product[] prodArray = new Product[10];
-            prodArray[0] = new Product(3.50, "Regular Coffee");
-            prodArray[1] = new Product(4.25, "Latte");
-            prodArray[2] = new Product(4.00, "Espresso");
-            prodArray[3] = new Product(3.25, "Cold Brew");
-            prodArray[4] = new Product(1.75, "Bagel");
-            prodArray[5] = new Product(2.25, "Muffin");
-            prodArray[6] = new Product(3.00, "Scone");
-            prodArray[7] = new Product(0.00, "Water");
-            prodArray[8] = new Product(1.00, "Tea");
-            prodArray[9] = new Product(12.00, "Logo Mug");
+            prodArray[0] = new Product(3.50m, "Regular Coffee");
+            prodArray[1] = new Product(4.25m, "Latte");
+            prodArray[2] = new Product(4.00m, "Espresso");
+            prodArray[3] = new Product(3.25m, "Cold Brew");
+            prodArray[4] = new Product(1.75m, "Bagel");
+            prodArray[5] = new Product(2.25m, "Muffin");
+            prodArray[6] = new Product(3.00m, "Scone");
+            prodArray[7] = new Product(0.00m, "Water");
+            prodArray[8] = new Product(1.00m, "Tea");
+            prodArray[9] = new Product(12.00m, "Logo Mug");
 
             return prodArray;
         }
@@ -33,7 +33,7 @@ namespace pointofsale
                 Console.WriteLine("Selection " + (i + 1) + ": " + ReturnArray[i].name.PadRight(15) + "        " + ReturnArray[i].price.ToString().PadLeft(10));
             }
         }
-        public double returnPrice(int i)
+        public decimal returnPrice(int i)
         {
 
             return prodArray[i].price;
